+++ Sources/ProjectDescription/ResourceFileElement.swift
@@ -1,19 +1,21 @@
 import Foundation
 
-/// File element
+/// Resource file element
 ///
 /// - glob: a glob pattern for files to include
 /// - folderReference: a single path to a directory
 ///
 /// Note: For convenience, an element can be represented as a string literal
-///       `"some/pattern/**"` is the equivalent of `FileElement.glob(pattern: "some/pattern/**")`
-public enum FileElement: Codable, Equatable {
+///       `"some/pattern/**"` is the equivalent of `ResourceFileElement.glob(pattern: "some/pattern/**")`
+public enum ResourceFileElement: Codable, Equatable {
     /// A glob pattern of files to include
-    case glob(pattern: Path)
+    /// and ODR tags
+    case glob(pattern: Path, tags: [String] = [])
 
     /// Relative path to a directory to include
     /// as a folder reference
-    case folderReference(path: Path)
+    /// and ODR tags
+    case folderReference(path: Path, tags: [String] = [])
 
     private enum TypeName: String, Codable {
         case glob
@@ -33,18 +35,20 @@ public enum FileElement: Codable, Equatable {
         case type
         case pattern
         case path
+        case tags
     }
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         let type = try container.decode(TypeName.self, forKey: .type)
+        let tags = try? container.decode([String].self, forKey: .tags)
         switch type {
         case .glob:
             let pattern = try container.decode(Path.self, forKey: .pattern)
-            self = .glob(pattern: pattern)
+            self = .glob(pattern: pattern, tags: tags ?? [])
         case .folderReference:
             let path = try container.decode(Path.self, forKey: .path)
-            self = .folderReference(path: path)
+            self = .folderReference(path: path, tags: tags ?? [])
         }
     }
 
@@ -52,32 +56,18 @@ public enum FileElement: Codable, Equatable {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(typeName, forKey: .type)
         switch self {
-        case let .glob(pattern: pattern):
+        case let .glob(pattern: pattern, tags: tags):
             try container.encode(pattern, forKey: .pattern)
-        case let .folderReference(path: path):
+            try container.encode(tags, forKey: .tags)
+        case let .folderReference(path: path, tags: tags):
             try container.encode(path, forKey: .path)
+            try container.encode(tags, forKey: .tags)
         }
     }
 }
 
-extension FileElement: ExpressibleByStringInterpolation {
+extension ResourceFileElement: ExpressibleByStringInterpolation {
     public init(stringLiteral value: String) {
         self = .glob(pattern: Path(value))
-    }
-}
-
-extension Array: ExpressibleByUnicodeScalarLiteral where Element == FileElement {
-    public typealias UnicodeScalarLiteralType = String
-}
-
-extension Array: ExpressibleByExtendedGraphemeClusterLiteral where Element == FileElement {
-    public typealias ExtendedGraphemeClusterLiteralType = String
-}
-
-extension Array: ExpressibleByStringLiteral where Element == FileElement {
-    public typealias StringLiteralType = String
-
-    public init(stringLiteral value: String) {
-        self = [.glob(pattern: Path(value))]
     }
 }