+++ Sources/TuistCache/ContentHashing/ResourcesContentHasher.swift
@@ -3,7 +3,7 @@ import TuistCore
 import TuistGraph
 
 public protocol ResourcesContentHashing {
-    func hash(resources: [FileElement]) throws -> String
+    func hash(resources: [ResourceFileElement]) throws -> String
 }
 
 /// `ResourcesContentHasher`
@@ -19,7 +19,7 @@ public final class ResourcesContentHasher: ResourcesContentHashing {
 
     // MARK: - ResourcesContentHashing
 
-    public func hash(resources: [FileElement]) throws -> String {
+    public func hash(resources: [ResourceFileElement]) throws -> String {
         let hashes = try resources
             .sorted(by: { $0.path < $1.path })
             .map { try contentHasher.hash(path: $0.path) }