+++ Sources/TuistKit/Commands/Cache/CacheWarmCommand.swift
@@ -7,6 +7,7 @@ import TuistSupport
 struct CacheWarmCommand: ParsableCommand {
     static var configuration: CommandConfiguration {
         CommandConfiguration(commandName: "warm",
+                             _superCommandName: "cache",
                              abstract: "Warms the local and remote cache.")
     }