+++ Sources/TuistKit/Commands/Scale/CloudLogoutCommand.swift
@@ -5,6 +5,7 @@ import TSCBasic
 struct CloudLogoutCommand: ParsableCommand {
     static var configuration: CommandConfiguration {
         CommandConfiguration(commandName: "logout",
+                             _superCommandName: "cloud",
                              abstract: "Removes any existing session to authenticate on the server with the URL defined in the Config.swift file")
     }