+++ Sources/TuistKit/Commands/Signing/EncryptCommand.swift
@@ -5,6 +5,7 @@ import TSCBasic
 struct EncryptCommand: ParsableCommand {
     static var configuration: CommandConfiguration {
         CommandConfiguration(commandName: "encrypt",
+                             _superCommandName: "signing",
                              abstract: "Encrypts all files in Tuist/Signing directory")
     }